Crate futures_await_async_macro

Procedural macro for the #[async] attribute.

This crate is an implementation of the #[async] attribute as a procedural macro. This is nightly-only for now as it's using the unstable features of procedural macros. Furthermore it's generating code that's using a new keyword, yield, and a new construct, generators, both of which are also unstable.

Currently this crate depends on syn and quote to do all the heavy lifting, this is just a very small shim around creating a closure/future out of a generator.

Functions

async

async_block

async_stream

async_stream_block